+++ lib/src/codecs/codecs.dart
@@ -18,3 +18,4 @@ export 'g_year_month/year_month_codec.dart';
 export 'g_month/gregorian_month_codec.dart';
 export 'unsigned_short/unsigned_short_codec.dart';
 export 'positive_integer/positive_integer_codec.dart';
+export 'non_positive_integer/non_positive_integer_codec.dart';