+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - Implement `xsd:gMonth` codec and data type
 - Implement `xsd:unsignedShort` data type codec
 - Implement `xsd:positiveInteger` data type codec
+- Implement `xsd:nonPositiveInteger` data type codec